Remove #10 NewItem 없이도 Create 가능할 것 같아서 삭제
